+++ CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.4.6] - 2025-09-06
+- Fix detection of `file.associations`
+- Add additional setting to declare extensions for use with other IDEs 
+
 ## [1.4.5] - 2025-09-05
 - Fix end ranges of scopes closed at EOF